requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Signature Verifier — requirements.txt
# GPU machine: 8GB VRAM, CUDA 12.1
#
# IMPORTANT — PyTorch GPU install:
#   pip install torch==2.4.0 torchvision==0.19.0 \
#       --index-url https://download.pytorch.org/whl/cu121
#   (Do NOT install torch from this file — use the command above)
#
# FAISS GPU:
#   pip install faiss-gpu==1.7.4
#   (Not faiss-cpu — the GPU version is ~50x faster for large indexes)
# ─────────────────────────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
# uvloop==0.20.0
httptools==0.6.4
python-multipart==0.0.9

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python-headless==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0

# ── Machine Learning ──────────────────────────────────────────────────────────
torch==2.6.0
torchvision==0.21.0
scikit-learn==1.5.2
numpy==1.26.4

# ── Vector Search ─────────────────────────────────────────────────────────────
# faiss-gpu is installed separately (see above)
# faiss-cpu==1.8.0   ← use only if no GPU available

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2

# ── Validation & Config ───────────────────────────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]==2.8.2

# ── Logging ───────────────────────────────────────────────────────────────────
loguru==0.7.2

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.2

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.37.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8

# ── Dev Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
